Order and reclamation tables with their controllers

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\article;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ArticleController extends Controller
 {
@@ -22,14 +23,21 @@ class ArticleController extends Controller
 
     public function addArticle(Request $request){
         try {
-            $article = article::create([
-                'designationArticle' => $request->designationArticle
-            ]) ;
-            return response()->json([
-                'articles added succesfully !' ,
-                $article->id
+            $authUser = Auth::user() ;
+            if($authUser->role == 'magasinier'){
 
-            ]) ;
+                $article = article::create([
+                    'designationArticle' => $request->designationArticle
+                ]) ;
+                return response()->json([
+                    'articles added succesfully !' ,
+                    $article->id
+
+                ]) ;
+            }
+            return response()->json([
+                'error' => 'Only magasinier can add articles'
+            ],400) ;
         } catch (Exception $exp) {
             return response()->json([
                 'error' => $exp->getMessage()
@@ -39,13 +47,19 @@ class ArticleController extends Controller
 
     public function editActicle(Request $request){
         try {
-
-            $article = article::find($request->id)->update([
-                'designationArticle' => $request->designationArticle
-            ]) ;
+            $authUser = Auth::user() ;
+            if($authUser->role == 'magasinier'){
+                $article = article::find($request->id)->update([
+                    'designationArticle' => $request->designationArticle
+                ]) ;
+                return response()->json([
+                    'message' => 'article updated succesfully !' ,
+                ]) ;
+            }
             return response()->json([
-                'message' => 'article updated succesfully !' ,
-            ]) ;
+                'error' => 'Only magasinier can do this action'
+            ],400) ;
+
         } catch (Exception $exp) {
             return response()->json([
                 'error' => $exp->getMessage()
@@ -55,10 +69,16 @@ class ArticleController extends Controller
 
     public function deleteArticle(Request $request){
         try {
-            article::find($request->id)->delete() ;
+            $authUser = Auth::user() ;
+            if($authUser->role == 'magasinier'){
+                article::find($request->id)->delete() ;
+                return response()->json([
+                    'success' => 'article deleted succesfully !'
+                ]) ;
+            }
             return response()->json([
-                'article deleted succesfully !'
-            ]) ;
+                'error' => 'Only magasinier can add articles'
+            ],400) ;
         } catch (Exception $exp) {
             return response()->json([
                 'error' => $exp->getMessage()
